chore(refactor): extract `cookies` into separate module
